requirements.txt
-e .
kafka-python>=2.0.0
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.10.0
pandera>=0.18.0
optuna-integration
optuna==3.5.0
lightning==2.2.1
seaborn
python-box
ensure
torch>=2.0.0
pytorch-lightning>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
mlflow>=2.5.0
hydra-core>=1.3.0
fastapi>=0.100.0
uvicorn>=0.22.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0
joblib>=1.3.0
tqdm>=4.65.0
pydantic>=2.0.0
python-multipart>=0.0.6